fix: 重构

+++ src/main/java/com/yumefusaka/SmartAirport/pojo/Entity/Luggage.java
@@ -21,7 +21,7 @@ public class Luggage {
     private String description;
 
     @TableField(value = "passenger_id")
-    private int passenger_id;
+    private long passenger_id;
 
     @TableField(value = "number")
     private int number;
