+++ src/main/java/com/yumefusaka/SmartAirport/pojo/VO/FlightVO.java
@@ -7,13 +7,13 @@ import java.time.LocalDateTime;
 @Data
 public class FlightVO {
 
-    private int id;
+    private Long id;
 
     private Long airline_id;
 
     private String flight_number;
 
-    private int capacity;
+    private Long capacity;
 
     private String departure_city;
